[Refactor] [1/N] to simplify the vLLM serving architecture
